Moving files related to metrics to metrics/ and rename references
